fixes bug with infinite while loop because I forgot to make a value an into in sweater_pattern_data

+++ sweater_pattern_data.py
@@ -16,7 +16,7 @@ def define_sleeve_variables(swatch_data, measurements):
     sleeve_variables['sleeve_total_increase'] = sleeve_variables['estimated_sleeve_stitches_at_top'] - sleeve_variables['stitches_at_wrist']
     sleeve_variables['sleeve_stitch_increase'] = 2
     sleeve_variables['sleeve_rows_between_increase'] = int((sleeve_variables['sleeve_rows'] / (sleeve_variables['sleeve_total_increase'] / sleeve_variables['sleeve_stitch_increase'])))
-    sleeve_variables['sleeve_total_num_increases'] = sleeve_variables['sleeve_total_increase'] / sleeve_variables['sleeve_stitch_increase']
+    sleeve_variables['sleeve_total_num_increases'] = int(sleeve_variables['sleeve_total_increase'] / sleeve_variables['sleeve_stitch_increase'])
     sleeve_variables['calculated_sleeve_stitches_at_top'] = sleeve_variables['stitches_at_wrist'] + (sleeve_variables['sleeve_stitch_increase'] * math.floor(sleeve_variables['sleeve_total_num_increases']))
